+++ packages/vsce/src/commands/disableCommands/index.ts
@@ -14,4 +14,5 @@ export * from "./disableLocalFileCommand";
 export * from "./disableProgramCommand";
 export * from "./disableTransactionCommand";
 export * from "./disableLibraryCommand";
+export * from "./disableJVMServerCommand";