+++ packages/vsce/src/commands/enableCommands/index.ts
@@ -14,4 +14,4 @@ export * from "./enableLocalFileCommand";
 export * from "./enableProgramCommand";
 export * from "./enableTransactionCommand";
 export * from "./enableLibraryCommand";
-
+export * from "./enableJVMServerCommand";